+++ CalculationHelper/ZoomBlt.cpp
@@ -5,34 +5,34 @@
 
 void CZoomBlt::DrawFullTrans(CDC* pDC, CDC* tmpdc, int x, int y, int w, int h, int sx, int sy, int sw, int sh, int col)
 {
-	double x2 = x * fx() / 2;
-	double y2 = y * fy() / 2;
-	double w2 = w * fx() / 2;
-	double h2 = h * fy() / 2;
+	int x2 = (int)(x * fx() / 2);
+	int y2 = (int)(y * fy() / 2);
+	int w2 = (int)(w * fx() / 2);
+	int h2 = (int)(h * fy() / 2);
 	pDC->TransparentBlt(x2, y2, w2, h2, tmpdc, sx, sy, sw, sh, col);
 }
 
 void CZoomBlt::DrawFullStd(CDC* pDC, CDC* tmpdc, int x, int y, int w, int h, int sx, int sy)
 {
-	double x2 = x * fx() / 2;
-	double y2 = y * fy() / 2;
-	double w2 = w * fx() / 2;
-	double h2 = h * fy() / 2;
+	int x2 = (int)(x * fx() / 2);
+	int y2 = (int)(y * fy() / 2);
+	int w2 = (int)(w * fx() / 2);
+	int h2 = (int)(h * fy() / 2);
 	pDC->BitBlt(x2, y2, w2, h2, tmpdc, sx, sy, SRCCOPY);
 }
 
 void CZoomBlt::DrawFullStretch(CDC* pDC, CDC* tmpdc, int x, int y, int w, int h, int sx, int sy, int sw, int sh)
 {
-	double x2 = x * fx() / 2;
-	double y2 = y * fy() / 2;
-	double w2 = w * fx() / 2;
-	double h2 = h * fy() / 2;
+	int x2 = (int)(x * fx() / 2);
+	int y2 = (int)(y * fy() / 2);
+	int w2 = (int)(w * fx() / 2);
+	int h2 = (int)(h * fy() / 2);
 	pDC->StretchBlt(x2, y2, w2, h2, tmpdc, sx, sy, sw, sh, SRCCOPY);
 }
 
 void CZoomBlt::DrawFullRect(CDC* pDC, int x, int y, int r, int b, int c)
 {
-	pDC->FillSolidRect(CRect(x * fx() / 2, y * fy() / 2, r * fx(), b * fy()), c);
+	pDC->FillSolidRect(CRect((int)(x * fx() / 2), (int)(y * fy() / 2), (int)(r * fx()), (int)(b * fy())), c);
 }
 
 RECT CZoomBlt::ScreenRect()
